orchestrator/requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain Dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
langchain-experimental==0.0.47
langgraph==0.0.20

# LLM Providers
openai==1.6.1
anthropic==0.8.1
ollama==0.1.7

# MongoDB
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver

# Vector Store (for RAG)
chromadb==0.4.22
# Note: faiss-cpu removed due to compatibility issues
# If you need FAISS, install separately with: pip install faiss-cpu or use faiss-gpu

# Web Search & Scraping
duckduckgo-search==4.1.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1

# API Framework
fastapi==0.108.0
uvicorn==0.25.0
websockets==12.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
tiktoken==0.5.2

# Utilities
python-multipart==0.0.6
httpx==0.26.0
tenacity==8.2.3  # For retry logic
colorama==0.4.6  # For colored terminal output

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Code Analysis Tools (for code agent)
black==23.12.1
pylint==3.0.3
ast-comments==1.2.2

# Alternative vector store options
sentence-transformers==2.2.2  # For embeddings if not using OpenAI
